clean testAvatar.js
